refactor: input types
