+++ src/pages/form/type.tsx
@@ -1,0 +1,8 @@
+export type inputValuetype = {
+    name: string;
+    date: string;
+    continent: string;
+    gender: string;
+    agreement: boolean;
+    image: string;
+  };
